move tuple domain predicate to pixels-presto.
