+++ pixels-presto/src/main/java/io/pixelsdb/pixels/presto/PixelsRecordCursor.java
@@ -34,12 +34,12 @@ import io.pixelsdb.pixels.core.PixelsReader;
 import io.pixelsdb.pixels.core.PixelsReaderImpl;
 import io.pixelsdb.pixels.core.TypeDescription;
 import io.pixelsdb.pixels.core.predicate.PixelsPredicate;
-import io.pixelsdb.pixels.core.predicate.TupleDomainPixelsPredicate;
 import io.pixelsdb.pixels.core.reader.PixelsReaderOption;
 import io.pixelsdb.pixels.core.reader.PixelsRecordReader;
 import io.pixelsdb.pixels.core.vector.*;
 import io.pixelsdb.pixels.presto.exception.PixelsErrorCode;
 import io.pixelsdb.pixels.presto.impl.PixelsPrestoConfig;
+import io.pixelsdb.pixels.presto.impl.PixelsTupleDomainPredicate;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -133,7 +133,7 @@ public class PixelsRecordCursor implements RecordCursor
         {
             domains = split.getConstraint().getDomains().get();
         }
-        List<TupleDomainPixelsPredicate.ColumnReference<PixelsColumnHandle>> columnReferences =
+        List<PixelsTupleDomainPredicate.ColumnReference<PixelsColumnHandle>> columnReferences =
                 new ArrayList<>(domains.size());
         for (Map.Entry<PixelsColumnHandle, Domain> entry : domains.entrySet())
         {
@@ -141,12 +141,12 @@ public class PixelsRecordCursor implements RecordCursor
             String columnName = column.getColumnName();
             int columnOrdinal = split.getOrder().indexOf(columnName);
             columnReferences.add(
-                    new TupleDomainPixelsPredicate.ColumnReference<>(
+                    new PixelsTupleDomainPredicate.ColumnReference<>(
                             column,
                             columnOrdinal,
                             column.getColumnType()));
         }
-        PixelsPredicate predicate = new TupleDomainPixelsPredicate<>(split.getConstraint(), columnReferences);
+        PixelsPredicate predicate = new PixelsTupleDomainPredicate<>(split.getConstraint(), columnReferences);
 
         this.option = new PixelsReaderOption();
         this.option.skipCorruptRecords(true);
